tidy up + new problems + fix zig zag problem
