+++ src/Challenges/Maths.cs
@@ -1,9 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using LeetCode.Model;
 
-namespace LeetCode.Easy
+namespace LeetCode
 {
-    public class Maths
+    public static class Maths
     {
         /// <summary>
         /// LeetCode problem 461 - Hamming Distance: https://leetcode.com/problems/hamming-distance/#/description
@@ -60,7 +64,6 @@ namespace LeetCode.Easy
                         if (j != gridHeight - 1) if (grid[i, j + 1] == 1) neighbourCount++;
                         if (i != gridWidth - 1) if (grid[i + 1, j] == 1) neighbourCount++;
                     }
-
                 }
             }
 
@@ -101,7 +104,8 @@ namespace LeetCode.Easy
             {
                 if (nums[i] == 1) current++;
                 else current = 0;
-                max = current > max ? current : max;
+                // max = current > max ? current : max;
+                max = Math.Max(current, max);
             }
             return max;
         }
@@ -235,6 +239,129 @@ namespace LeetCode.Easy
             // Sister gets half of all candies - if less than half are unique, then the max unique to be gained is the total unique.
             if (uniqueCandies.Count < candies.Length / 2) return uniqueCandies.Count;
             else { return candies.Length / 2; } // If more than half are unique, max unique to be gained is half the total.
+        }
+
+        public static bool JudgeCircle(string moves)
+        {
+            int up = 0, across = 0;
+            foreach (char c in moves)
+            {
+                if (c == 'U') up++;
+                if (c == 'D') up--;
+                if (c == 'L') across++;
+                if (c == 'R') across--;
+            }
+            return (up == 0 && across == 0);
+        }
+
+        public static IList<int> SelfDividingNumbers(int left, int right)
+        {
+            var result = new List<int>();
+            for (int i = left; i <= right; i++)
+            {
+                var intString = Convert.ToString(i);
+                bool SelfDividing = true;
+                foreach (char c in intString)
+                {
+                    var charInt = (int)Char.GetNumericValue(c);
+                    if (charInt == 0) SelfDividing = false;
+                    else if (i % charInt != 0) SelfDividing = false;
+                }
+                if (SelfDividing) result.Add(i);
+            }
+            return result;
+        }
+
+
+        /// <summary>
+        /// LeetCode problem 537 - Complex Number Manipulation: https://leetcode.com/problems/complex-number-multiplication/#/description
+        /// </summary>
+        /// <param name="a">Complex number 1 - as a string.</param>
+        /// <param name="b">Complex number 2 - as a string.</param>
+        /// <returns>Complex number - 1 * 2 - as a string.</returns>
+        public static string ComplexNumberMultiply(string a, string b)
+        {
+            // Split string into real and imaginary components.
+            string[] leftStrings = a.Split('+');
+            string[] rightStrings = b.Split('+');
+
+            // Convert strings into integers:
+            int leftReal = Convert.ToInt32(leftStrings[0]);
+            int leftImaginary = Convert.ToInt32(leftStrings[1].TrimEnd('i'));
+            int rightReal = Convert.ToInt32(rightStrings[0]);
+            int rightImaginary = Convert.ToInt32(rightStrings[1].TrimEnd('i'));
+
+            // Final real number:
+            int finalReal = (leftReal * rightReal) - (leftImaginary * rightImaginary);
+            // Final imaginary number:
+            int finalImaginary = (leftReal * rightImaginary) + (leftImaginary * rightReal);
+
+            return (finalReal.ToString() + "+" + finalImaginary.ToString() + "i");
+        }
+
+        /// <summary>
+        /// LeetCode problem 2 - Add Two Numbers: https://leetcode.com/problems/add-two-numbers/#/description
+        /// </summary>
+        /// <param name="l1">First number.</param>
+        /// <param name="l2">Second number.</param>
+        /// <returns>Sum of 2 inputs.</returns>
+        public static ListNode AddTwoNumbers(ListNode l1, ListNode l2)
+        {
+            ListNode result;
+
+            if (l1 == null)
+            {
+                if (l2 == null) return null;
+                else return l2;
+            }
+            else if (l2 == null) return l1;
+
+            // L1 and L2 are the smallest numbers in a sequence.
+            // e.g if L1 == 2, and links to another ListNode 1, which doesn't link to anything else, this represents the number 12.
+
+            // Add up this node:
+            int value = l1.val + l2.val;
+
+            // If an extra 1 needs to be carried to the next sum:
+            int carry = 0;
+            if (value > 9) { carry = 1; value -= 10; }
+
+            result = new ListNode(value);
+            result.next = AddTwoNumbers(l1.next, l2.next);
+            if (result.next == null && carry != 0) result.next = new ListNode(carry);
+            else if (result.next != null) result.next = AddTwoNumbers(result.next, new ListNode(carry));
+            return result;
+        }
+
+        // https://leetcode.com/problems/max-increase-to-keep-city-skyline/description/
+        public static int MaxIncreaseKeepingSkyline(int[][] grid)
+        {
+            var topView = new int[grid.Length];
+            var sideView = new int[grid.Length];
+            int result = 0;
+
+            for (int i = 0; i < grid.Length; i++)
+            {
+                sideView[i] = grid[i].Max();
+                int largestVertical = 0;
+                for (int j = 0; j < grid.Length; j++)
+                {
+                    largestVertical = Math.Max(grid[j][i], largestVertical);
+                }
+                topView[i] = largestVertical;
+            }
+
+            for (int i = 0; i < grid.Length; i++)
+            {
+                for (int j = 0; j < grid.Length; j++)
+                {
+                    var currentHeight = grid[i][j];
+                    var difference = Math.Min(sideView[i], topView[j]) - currentHeight;
+                    if (difference > 0) result += difference;
+                }
+            }
+
+            return result;
         }
     }
 }
