+++ src/Tests/StringTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Xunit;
-using static LeetCode.Easy.String;
+using static LeetCode.String;
 
 namespace LeetCode.Easy.Tests
 {
@@ -18,7 +18,7 @@ namespace LeetCode.Easy.Tests
             [InlineData("111", "111")]
             public void ValidInputs(string input, string expectedResult)
             {
-                var actualResult = LeetCode.Easy.String.ReverseString(input);
+                var actualResult = LeetCode.String.ReverseString(input);
                 Assert.Equal(expectedResult, actualResult);
             }
         }
@@ -96,6 +96,63 @@ namespace LeetCode.Easy.Tests
                 string actualResult = ReverseWordsInString(input);
                 Assert.Equal(expectedResult, actualResult);
             }
+        }
+
+        public class NumJewels
+        {
+            [Theory]
+            [InlineData("aA", "aAAbbbb", 3)]
+            [InlineData("z", "ZZ", 0)]
+            public void CountNumJewels(string J, string S, int expectedResult)
+            {
+                var actualResult = String.NumJewelsInStones(J, S);
+                Assert.Equal(expectedResult, actualResult);
+            }
+        }
+        public class LongestNonRepeatingSubString
+        {
+            [Theory]
+            [InlineData("aaa", 1)]
+            [InlineData("aab", 2)]
+            [InlineData("abc", 3)]
+            [InlineData("dvdf", 3)]
+            [InlineData("abcabcbb", 3)]
+            [InlineData("asjrgapa", 6)]
+            [InlineData("jbpnbwwd", 4)]
+            public void ValidInputs(string x, int expectedResult)
+            {
+                int actualResult = LengthOfLongestSubstring(x);
+                Assert.Equal(expectedResult, actualResult);
+            }
+        }
+
+        public class LongestPalindromicSubstring
+        {
+            [Theory]
+            [InlineData("aba", "aba")]
+            [InlineData("abb", "bb")]
+            [InlineData("babad", "bab")]
+            [InlineData("aaaaaaabbbbbbb", "aaaaaaa")]
+            [InlineData("aaaaaaabbbbbbbb", "bbbbbbbb")]
+            public void ValidInputs(string s, string expectedResult)
+            {
+                string actualResult = LongestPalindrome(s);
+                Assert.Equal(expectedResult, actualResult);
+            }
+        }
+
+        public class ZigZagConversion
+        {
+            [Theory]
+            [InlineData("PAYPALISHIRING", 1, "PAYPALISHIRING")]
+            [InlineData("PAYPALISHIRING", 3, "PAHNAPLSIIGYIR")]
+            [InlineData("PAYPALISHIRING", 4, "PINALSIGYAHRPI")]
+            public void ValidInputs(string s, int rows, string expectedResult)
+            {
+                string actualResult = ZigZagConvert(s, rows);
+                Assert.Equal(expectedResult, actualResult);
+            }
+
         }
     }
 }
